proxystore: stabilize order for duplicates in response losertree
Currently the order of responses in the losertree on collisions in the
labelset is random. This can happen if we drop the replica label in an
endpoint.
In the case of sidecars the order of responses has effect on
deduplication. The primary iterator is used until we find a large enough
gap to failover to the replica iterator, where primary and replica is
determined by the order they are returned from the proxy losertree.
This can lead to slight differences if we repeat a query since different
sidecars have scraped at different times possibly.
Using the store labelset as tiebreaker is an attempt at stabilizing
this.

Signed-off-by: Michael Hoffmann <mhoffmann@cloudflare.com>